requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pandas==2.0.3
pydantic==2.4.0
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
scikit-learn==1.3.2
sentence-transformers==2.2.2
huggingface-hub==0.19.4
python-dotenv==1.0.0
requests